nec2c-0.9.tar.bz2 import to git
  Version 0.9: After a bug report by David J. Singer, I changed all
  declarations of variables that are used in memory allocations, from int
  to size_t. This error was in the nec2c code from the beginning but
  apparently it only showed up when extremely large memory allocations
  are requested in nec2c.

+++ misc.c
@@ -180,8 +180,7 @@ int load_line( char *buff, FILE *pfile )
 /*------------------------------------------------------------------------*/
 
 /***  Memory allocation/freeing utils ***/
-
-void mem_alloc( void **ptr, int req )
+void mem_alloc( void **ptr, size_t req )
 {
   free_ptr( ptr );
   *ptr = malloc( req );
@@ -192,7 +191,7 @@ void mem_alloc( void **ptr, int req )
 
 /*------------------------------------------------------------------------*/
 
-void mem_realloc( void **ptr, int req )
+void mem_realloc( void **ptr, size_t req )
 {
   *ptr = realloc( *ptr, req );
   if( *ptr == NULL )
